+++ resources/views/auth/login.blade.php
@@ -1,4 +1,7 @@
 <x-guest-layout>
+    <x-slot name="titlePostfix">
+        {{__('Login')}}
+    </x-slot>
     <!-- Session Status -->
     <x-auth-session-status class="mb-4" :status="session('status')" />