+++ resources/views/layouts/app.blade.php
@@ -4,8 +4,7 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <meta name="csrf-token" content="{{ csrf_token() }}">
-
-    <title>{{ \App\Models\Configuration::getString(\App\Models\ConfigurationKey::APPEARANCE_APP_NAME) ?: config('app.name') }}</title>
+    <title>{{ $title }}</title>
 
     <!-- Scripts -->
     @vite(['resources/css/app.css', 'resources/js/app.js'])